Implement Elog; add a Texter interface
Uses a slice of ElogOperations instead of the typeless buffer
in the C version.  Forgoes all the flush logic in favor of managing
explicit additions.

Re-uses allocations as far as possible to avoid making garbage.

Does not coallesce Replace operations - they seem rare.  Can implement
if we need them.

Texter interface exposes the operations done on a Text from elog,
allowing implementation of a test harness.  Probably this object should
be called a "Selection Buffer", as it manages a Buffer with a Selection.

The TextMock implementation in elog_test.go is probably a fullish
buff.go implementation.