[frame] Start rationalizing the frame accessors
The C version of the Frame object has a number of fields that can be
read and sometimes written from outside. This is an unfortunate
encapsulation violation and can make it harder to reason about the
code. Start formalizing which of these can be private and which public
and provide accessors for some of the existing private fields that are
obviously used by an Acme text instance.