[frame/framedemo] working selection and editing
A series of changes to accomplish the following:

* get Drawsel0 to correctly draw a specified selection highlight if
none is currently applied.

* get Drawsel to correctly update both the tick and selection
highlight regardless of the state of the tick or selection highlight.

* extend the framedemo to show that selections work properly in both
directions.

This version of Drawsel may not be semantically equivalent to the one
found in the C code. In particular: the way that selection range is
both used and adjusted from outside and inside of Frame is undesirable
in a Go context so I attempted to alter the semantics in a reasonable
way.