[acme] Font rationalization
Recall that in Plan9, the font name is a path that corresponds to the
DIP name of the font. The Dump command (#12) needs to create dump
files that use DIP font names to permit loading correctly on a display
with a different high DPI setting.

draw.Font will (transparently) allocate an appropriate physical pixel
font corresponding to the specified DIP font name based on whether or
not the backing Display is high-DPI or low-DPI and its Name instance
variable is the physical font. Consequently, Dump cannot create
correct dump files by the use of draw.Font.Name.

Instead, always use the string DIP name of the font in Edwood
throughout and retrieve the correct metrics or backing draw.Font
object on an as-needed basis based on the font DIP name.