[frame] Remove nalloc, nbox
There have been several frame bugs related to the fact that Frame.box
is a slice but with manual bounds management via Frame.nalloc and
Frame.nbox. As a result, it was difficult and subtle to use Go
constructs like append on the slice of frbox objects.

Make this code more idiomatic and hopefully less buggy (in pursuit of
great robustness per #30) by making Frame.box a true Go slice.