Alter semantics of Select
Frame.Select calls a provided callback to get more lines of text
during a scroll operation.  This is at odds with adding locks to
Frame as Frame.Select must acquire (and hold) a lock for the duration
of its operation to prevent mutators of the Frame while dragging.
Instead, have the callback provide an alternate Frame instance that
bypasses lock acquisition.