[frame] Fix handling of delete with tabs
Frame was incorrectly updating the width of a tab box while editing
and hence causing #59 to happen. Update the width of the tab box
later during the execution of the delete code.

Also add a convenient tool testedwood.sh to run Edwood in a test
configuration so that it's easier to test Edwood while editing the code
in Edwood.