Shutdown connections from forked children (#96)
Edwood child processes forked to run |command were lingering and
not closing the Acme filesystem connection back to the parent Edwood
until (sometimes) significantly later. When the child xfidclose-ed
the wrsel file, it would change the parent's selection back to what
it was when command was first middle-clicked. This caused issue
soon as command completes. It also reverts part of ffe7178 now that
the underlying cause of seeing the panic (asynchronous selection
mutation) has been fixed.