Short term mitigation for column blackness #52 (#129)
* code cleanup in col.go

Start digging into the column handling code. Helps #52. Code
structuring for easier understanding.

* Remove unused code from col.go

Delete some unused code from col.go after some previous code cleanup.

* col.go speculative bug fixes

Some speculative bug fixes for Column.

* Additional trivial code cleanup in col.go

* col.go temporary mitigation for #52

While investigating #52, I noted that the problem is caused when we
try to resize a Window object to 0 lines of body (with an implied
collapsing of the tag.) This change mitigates the bug in #52 by not
letting us get to 0 lines of body. Helps #52. A better change is
forthcoming.

* Ran gofmt

* Ran gofmt correctly.