Reimplement regular expression
Edwood requires a regular expression implementation that works with
runes, correctly handles ^ and $ for selection, and enables backward
search. This commits adds an internal package `internal/regexp` that is
a copy of Go's regexp package but adds the methods Edwood needs (see the
README file in the package directory).

Also included are changes that make Edwood use the new regular
expression package. The Texter.View method is changed to return []rune
instead of []byte. This is harmless because this method was not being
used anywhere.

Fixes #157