Reduce use of File's internal details (#195)
* Relocate the typing cache

Text maintains a cache of text insertions that should be Undo-able
as a group. Relocate this cache into File so that the File API surface
makes a small step closer to the undo.Buffer which encapsulates
undo with well-defined Commit points.

A step towards #97.

* Modify Text to use file.ReadC directly.

* Additional code cleanup in Text

Post relocating of the typing cache, cleanup some additional code.

* Stop looking into File's internals for the buffer size.

There is a frequent pattern in the code to reach into File's buffer
implementation to determine the number of runes contained in the
File. This impedes replacing File's internal details so switch
external users to the Size function on File objects.

Helps with #97.

* Align File.Commit with undo.Buffer

Modify File.Commit to have API alignment with undo.Buffer's Commit.
Helps with #97.

* Rename Insert to InsertAt and Delete to DeleteAt

Advance the semantic fit between undo.Buffer and File by renaming the
existing Insert and Delete methods to InsertAt and DeleteAt.

Helps #97.

* Pull hash into separate file

Move the FileHash code into a separate file to simplify the File code. Cleanup
to help with #97.

* File.InsertAt runs its observers

Refactor the code in Text.Insert to let File.InsertAt (uniformly) invoke
an observer pattern on any insertion. Helps with #97.

NB: not complete. Rebase in follow-on cleanup CLs. In particular,
Undo doesn't redraw the box correctly in this change.

* File.DeleteAt runs all its observers

Refactor the code in DeleteAt to run observers. Helps #97. Fixes the
problem with deletion in Zerox Windows noted in the previous CL.

* Remove an unncessary use of File.text

Helps with #97. Moving towards to keeping the observer pattern an
opaque detail.

* Introduce File.HasMultipleTexts()

Reduce the number of places that look into internal details of File by
adding a HasMultipleTexts accessor. Simplifies code for helping #97.

* Cleanup curtext use

Altering File for #97 permits removing the use of curtext from allupdate.

* Trivial code cleanup

* Convert File.Load to use File.InsertAt

Rewrite File.Load in terms of File.InsertAt to adapt it to the
undo.Buffer API surface. This also simplifies the code by Re-using the
new observer structure. Helps with #97.

* Cleanup File.Close

File.Close does not map to undo.Buffer so clean up this and some other
unnecessary code. Helps with #97.

* Remove unnecessary File.text access

Remove an unnecessary use of File's internal details. Helps #97.

* Trivial code cleanup

Remove some unnecessary code and add a comment.

* Assorted code cleanups in Window

Window shouldn't reach into File internal details. Start reducing this.
Code cleanup tangential to #97.

* Minor code cleanups.

Further reduce the cases where Window reaches into File's deails.

* Fix a rebase edit error