Fix crash when snarfing empty selection using mouse chord
The issue was introduced during transliteration of the original C code.
Acme's `fileundo` function takes pointers to `q0` and `q1` as argument,
so there was no need for it to indicate if `q0` or `q1` was changed.
Since we return `q0` and `q1` in the Go code, we need to indicate if
caller should make use the new values.

Fixes #230