Fix race when saving to dump file before exit
When Edwood closes connection to the file server before exit, wait for
`9pserve` can return this error:

	9pserve wait failed: waitid: no child processes.

It's important that we don't panic for this error, so that Edwood has a
chance to finish cleaning up before exit. In particular, if Edwood is in
the middle of writing to the dump file, the dump file may end up being
truncated to an empty file.

We also move `killprocs` to *after* `row.Dump`. This leads to a dump
file more consistent with the `Dump` command. It ensures that all the
external commands currently running show up in the row tag inside the
dump file, even if its undesirable (at least to me).

Also, do not write to dump file on `Exit` command. We only write to dump
file on unexpected signals (SIGINT, SIGTERM, etc.). This behavior is
consistent with acme.

Removed some unnecessary (and racy) synchronization using PID and global
variable `dumping`. The C code needed them because the signal handler was
a callback. We receive the signal from a channel.

The TODO regarding saving the state on crash is a separate issue, which
we can solve by not crashing in the first place. We should always try to
recover from error (and maybe show an warning) instead of panic. Also,
autosave (issue #71) can help here.