Fix xfidruneread for small reads and multi-byte characters
We were crashing if we were trying to read less than 4 (utf8.UTFMax)
runes. The function also didn't handle muti-byte characters correctly
and it returned the number of bytes read instead of the number of runes.

I've changed the function to be more similar to the C code, decoding
runes from the beginning of the bytes buffer instead of try to find the
rune boundary from the end of the buffer. Finding the boundary doesn't
solve the problem of computing the number of runes, which we need to
return.

Fixes #255