Test Text.DirName and fix plumbing from row tag
* When plumbing from the row tag, we were sending "" as the `Dir` in the
`plumb.Message`. We now send the absolute path to current working
directory instead.

* Remove `dirname` function, which had the same functionality as
`Text.DirName`.

* Change `Text.DirName` to always return a cleaned path (as returned by
`filepath.Clean`). This means we'll never return "", which becomes "."
after clean.

* Removed some hacks (marked with "sigh") where "." was being converted
to "" for some unknown reason. As far as I can see, this hack isn't
necessary in Go. When "." is used in `filepath.Join` and `exec.Cmd.Dir`,
it works as expected.