Fix crash when writing muti-line commands containing deletion to ctl file
* Fixes #272

* Stop processing input on error "file dirty". This behavior now matches
acme.

* Also stop processing on error "bad character in file name". This
behavior does not match acme but I don't see why we should make use of a
bad window name, and then continue processing more input (which can
overwrite the error).

* Simplify how we compute the number bytes that were processed
(`Fcall.Count` in response). Generally, we always process the full
buffer if we don't encounter an error. For error, we send a `Rerror`
response, which has no count.