Fix reconnection to plumber
The `9fans.net/go/plumb` package caches the connection (`*client.Fsys`)
to the plumber. So, if the plumber isn't running when Edwood starts up, we
will never gain a connection to the plumber, even if the plumber is
started. Similarly, we never regain connection after losing it. The fix
manually connects to the plumber using `client.MountService`.

Also fix some missed `Fid.Close`: plumbsendfid and plumbeditfid wasn't
being closed in some branches.