more exec tests (#366)
* work-in-progress CL

Start of a change that is intended to provide additional test coverage for
all of the commands in exec. Here, I'm writing a test for the Cut command.
It however requires some kind of mocking of the display object.

* Test of exec.go:/cut/

Start of a test for the cut function in exec.go.