Observer-driven tag updates
Use an observer scheme to invoke `Window.setTag` instead of the manual
calls to `Window.SetTag`. This significantly reduces the number of
calls to `SetTag`: e.g. in editing this commit message, we reduced the
number of `setTag` invocations by ~99%.

This change also contributes to #400 by setting up the start of an
observer structure for plumbing together Undo actions *in the tag*
into filename changes.