Add a better E
Plan9 port comes with a command E. E is a shell script that opens a
file for editing and blocks until one closes it in Acme/Edwood. P9P E
does not work for files whose names contain spaces. So this CL adds a
new E implementation in Go that should support file names containing
spaces. It doesn't yet do so because of bugs in Edwood.

Also updated go.mod for E commandt ake sure that the go.mod and go.sum
are correct for the deps needed.

Also updated the code for the new version of staticcheck.